requirements.txt
alabaster==0.7.12
asgiref==3.4.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dataclasses==0.6
Django==2.2
django-redis==4.11.0
djangorestframework==3.10.3
docutils==0.17.1
hyperlink==21.0.0
idna==2.10
imagesize==1.2.0
incremental==21.3.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
paho-mqtt==1.5.1
paramiko==2.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
qcloud-python-sts==3.0.5
qcloudsms-py==0.1.4
redis==3.5.3
requests==2.25.1
service-identity==21.1.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
telegram==0.0.1
tencentcloud-sdk-python==3.0.429
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
zope.interface==5.4.0
